adding maintenance items and making pep 8 compatible
